requirements.txt
streamlit
graphviz
antlr4-tools
antlr4-python3-runtime
pandas